Refactor static methods to separate classes
